requirements.txt
openai
uvicorn
python-dotenv
langchain
FAISS-gpu
sentence_transformers